renamed all files kebab-style and added unicorn to the linter
