+++ .eslintignore
@@ -1,0 +1,3 @@
+build
+coverage
+node_modules
