+++ package.json
@@ -32,6 +32,7 @@
     "eslint": "^8.41.0",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-unicorn": "^47.0.0",
     "gh-pages": "^5.0.0",
     "jest-axe": "^7.0.1",
     "prettier": "^2.8.8",
@@ -48,7 +49,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "test-cov": "jest --coverage",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "eslint ."
   },
   "jest": {
     "transform": {
