+++ src/App.test.tsx
@@ -1,7 +1,7 @@
 import { render } from '@testing-library/react';
-import App from './App';
+import App from './app';
 import { ThemeProvider } from '@mui/material/styles';
-import darkTheme from './themes/darkTheme';
+import darkTheme from './themes/dark.theme';
 import React from 'react';
 
 // Mock the theme provider
