+++ src/components/Footer.a11y.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { axe, toHaveNoViolations } from 'jest-axe';
-import Footer from '../components/Footer';
+import Footer from './footer';
 
 expect.extend(toHaveNoViolations);
