+++ src/components/Footer.test.tsx
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import Footer from './Footer';
+import Footer from './footer';
 import React from 'react';
 
 describe('Footer', () => {
