+++ src/components/Header.a11y.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import { axe, toHaveNoViolations } from 'jest-axe';
-import Header from '../components/Header';
+import Header from './header';
 
 expect.extend(toHaveNoViolations);
