+++ src/components/photo-grid.component.tsx
